requirements.txt
flask
pyodbc
requests
python-decouple
gunicorn
